// src/app/Models/Publicity.ts
export class Publicity {
  idPublicity: number;
  namePublicity: string;
  startDatePublicity: Date;

  endDatePublicity: Date;

  costPublicity: number;
  statusPublicity: boolean;
  mailOwnerPublicity: string;
  phoneOwnerPublicity: number;
  typePublicity: string;
  descriptionPublicity: string;
  InitialViewNumber: number;
  finalViewNumber: number;
  fileName: string;
}